Move some client side logic from ApiPort.exe to ApiPortClient
